+++ lib/workload/stateful/stacks/token-service/deploy/construct/policy/index.ts
