+++ lib/workload/stateful/stacks/token-service/token_service/rotate_service_user.py
